split mail schedule to its own service preparing for cleanup service
