+++ SendItems/Events/ModEvents.cs
@@ -7,7 +7,8 @@ namespace Denifia.Stardew.SendItems.Events
         public static event EventHandler<MailComposedEventArgs> MailComposed;
         public static event EventHandler PlayerUsingLetterbox;
         public static event EventHandler PlayerUsingPostbox;
-        public static event EventHandler OnMailDeliverySchedule;
+        public static event EventHandler OnMailDelivery;
+        public static event EventHandler OnMailCleanup;
         public static event EventHandler MailDelivered;
         public static event EventHandler<MailReadEventArgs> MailRead;
 
@@ -26,9 +27,14 @@ namespace Denifia.Stardew.SendItems.Events
             PlayerUsingPostbox?.Invoke(sender, e);
         }
 
-        internal static void RaiseOnMailDeliverySchedule(object sender, EventArgs e)
+        internal static void RaiseOnMailDelivery(object sender, EventArgs e)
         {
-            OnMailDeliverySchedule?.Invoke(sender, e);
+            OnMailDelivery?.Invoke(sender, e);
+        }
+
+        internal static void RaiseOnMailCleanup(object sender, EventArgs e)
+        {
+            OnMailCleanup?.Invoke(sender, e);
         }
 
         internal static void RaiseMailDelivered(object sender, EventArgs e)
